Backend/requirements.txt
# ===== DEPENDENCIAS PRINCIPALES =====

# Framework web asíncrono para crear la API REST
fastapi==0.104.1

# Servidor ASGI para ejecutar FastAPI
uvicorn[standard]==0.24.0

# Cliente oficial de OpenAI para comunicarse con GPT
openai==1.3.7

# Gestión de variables de entorno
python-dotenv==1.0.0

# Validación de datos y serialización JSON
pydantic==2.5.0

# Soporte para CORS (Cross-Origin Resource Sharing)
python-multipart==0.0.6

# ===== DEPENDENCIAS ADICIONALES =====

# Mejores tipos de datos para Python
typing-extensions==4.8.0

# Procesamiento de fechas y tiempo
python-dateutil==2.8.2

# Logging avanzado
loguru==0.7.2

# ===== DEPENDENCIAS DE DESARROLLO (OPCIONALES) =====

# Testing framework
pytest==7.4.3
pytest-asyncio==0.21.1

# Formateo de código
black==23.11.0

# Linting
flake8==6.1.0

# Documentación automática adicional
python-json-logger==2.0.7